Initial commit of the demo refactor.
